refactor: router.tsx에 routes 옮기기
